+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - add BoxDrawingBuilder().align(column, HorizontalAlignment)
 - add BoxDrawingBuilder().setDefaultPad(left_spaces, right_spaces)
+- add BoxDrawingBuilder().alignLeft(columns[])
 
 ## `1.1.0` 13 Jan 2023